+++ LatteView.php
@@ -7,32 +7,35 @@ use Latte\Macros\MacroSet;
  *
  * @package ProcessWire
  */
-class LatteView {
+class LatteView
+{
 
-    public function invokeFilter($name, $args) {
-
-        if (!is_array($args))
+    public function invokeFilter($name, $args)
+    {
+        if (!is_array($args)) {
             $args = array($args);
+        }
 
         return $this->latte->invokeFilter($name, $args);
     }
 
 
-    public function addMacro() {
-
+    public function addMacro()
+    {
         $args = func_get_args();
 
         $set = new MacroSet($this->latte->getCompiler());
 
-        if (!is_array($args))
+        if (!is_array($args)) {
             $args = array($args);
+        }
 
         call_user_func_array(array($set, 'addMacro'), $args);
     }
 
 
-    public function addFilter($name, $closure) {
+    public function addFilter($name, $closure)
+    {
         $this->latte->addFilter($name, $closure);
     }
-
 }